Migrate email sender to Flask CLI
