+++ config.py.template
@@ -8,5 +8,4 @@ MAIL_USE_SSL = True
 MAIL_USERNAME = "username"
 MAIL_PASSWORD = "password"
 MAIL_DEFAULT_SENDER = ("name", "address")
-PASSWORD_HASH = b""  # It can be generated with `python -c "import os; print(os.urandom(32))"`
-PASSWORD_SALT = b""  # It can be generated with `python -c "import hashlib; print(hashlib.pbkdf2_hmac("sha256", b"<password>", b"<salt>", 100000))"`
+SERVER_NAME = "127.0.0.1:5000"  # Needed to generate the HTML email
